content/canvas/test/test_2d.gradient.object.type.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.gradient.object.type</title>
<!-- Testing: window.CanvasGradient exists and has the right properties -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(window.CanvasGradient !== undefined, "window.CanvasGradient !== undefined");
ok(window.CanvasGradient.prototype.addColorStop !== undefined, "window.CanvasGradient.prototype.addColorStop !== undefined");

SimpleTest.finish();

});
</script>
